requirements.txt
flask>=2.0
picamera2
psutil
Pillow
opencv-python-headless
numpy
ffmpeg-python
pyyaml
piexif
